+++ app/src/main/kotlin/me/xizzhu/android/joshua/reading/detail/VerseDetailViewLayout.kt
@@ -77,13 +77,15 @@ class VerseDetailViewLayout : FrameLayout {
         })
     }
 
-    fun setListeners(onClicked: () -> Unit, onBookmarkClicked: () -> Unit, onHighlightClicked: () -> Unit,
-                     onNoteUpdated: (String) -> Unit, onNoStrongNumberClicked: () -> Unit) {
+    fun initialize(onClicked: () -> Boolean, onBookmarkClicked: () -> Unit, onHighlightClicked: () -> Unit,
+                   onNoteUpdated: (String) -> Unit, onNoStrongNumberClicked: () -> Unit) {
         setOnClickListener { onClicked() }
         bookmark.setOnClickListener { onBookmarkClicked() }
         highlight.setOnClickListener { onHighlightClicked() }
         adapter.onNoteUpdated = onNoteUpdated
         adapter.onNoStrongNumberClicked = onNoStrongNumberClicked
+
+        post { hide() }
     }
 
     fun setSettings(settings: Settings) {